Updates Move Types to new header.
